Reorganize in-flight log mechanism
Coordinate communication between upstream and downstream in order to
send downstream in-flight tuples first before new ones.
The process for processing an in-flight log becomes the following:
- downstream sends upstream a request to prepare to send the in-flight
log starting from a specific checkpoint
- upstream sends downstream an ack that it received the request
- upstream clears the buffers of the corresponding subpartition
connected with downstream
- upstream waits for a request to replay the in-flight log
- downstream sends a request to upstream to replay the in-flight log
after any state restore process completes
- downstream completes the in-flight log process (and input channel
re-set) when it receives the ack from upstream

The downstream part of the in-flight log implementation happens when
re-setting input channels.
Only when all input channels ar re-set does the standby task begin
running.
This is enforced in order to be certain that the main thread won't
request partitions and start processing records before the in-flight log
is set to run.